+++ tests/testthat/test-file-append.R
@@ -1,9 +1,10 @@
-context("file open")
+context("file append")
 
 library("httr")
 library("readr")
 library("digest")
 library("utils")
+library("dplyr")
 
 # see helper.R for setup
 
@@ -13,6 +14,15 @@ if (interactive()){
   filename_csv_original <- tempfile(fileext = ".csv")
   write_csv(iris, filename_csv_original)
 
+  filename_csv_append <- tempfile(fileext = ".csv")
+  write_csv(iris, filename_csv_append, col_names = FALSE)
+
+  filename_csv_double_original <- tempfile(fileext = ".csv")
+  write_csv(
+    bind_rows(iris, iris),
+    filename_csv_double_original
+  )
+
   adls_create(
     helper$adls,
     file = upload_file(filename_csv_original),
@@ -20,16 +30,11 @@ if (interactive()){
     overwrite = FALSE
   )
 
-  filename_zip_original <- tempfile(fileext = ".zip")
-  zip(filename_zip_original, filename_csv_original)
-
-  adls_create(
+  adls_append(
     helper$adls,
-    file = upload_file(filename_zip_original),
-    path = "iris.zip",
-    overwrite = FALSE
+    file = upload_file(filename_csv_append),
+    path = "iris.csv"
   )
-
 }
 
 test_that("the text is the same", {
@@ -38,7 +43,7 @@ test_that("the text is the same", {
   skip_on_travis()
   skip_if_not(interactive())
 
-  text_original <- read_file(filename_csv_original)
+  text_original <- read_file(filename_csv_double_original)
   text_test <- adls_open_to_text(
     helper$adls,
     path = "iris.csv"
@@ -63,36 +68,13 @@ test_that("the file is the same", {
   )
 
   expect_identical(
-    digest(filename_csv_original, file = TRUE),
+    digest(filename_csv_double_original, file = TRUE),
     digest(filename_csv_test, file = TRUE)
-  )
-
-})
-
-test_that("a zip file is the same", {
-
-  skip_on_cran()
-  skip_on_travis()
-  skip_if_not(interactive())
-
-  filename_zip_test <- tempfile(fileext = ".zip")
-  text_test <- adls_open_to_file(
-    helper$adls,
-    path = "iris.zip",
-    path_local = filename_zip_test
-  )
-
-  expect_identical(
-    digest(filename_zip_original, file = TRUE),
-    digest(filename_zip_test, file = TRUE)
   )
 
 })
 
 # delete file from data lake
 if (interactive()){
-
   adls_delete(helper$adls, path = "iris.csv")
-  adls_delete(helper$adls, path = "iris.zip")
-
 }