+++ NEWS.md
@@ -1,8 +1,10 @@
 # AzureDatalakeStore 0.0.0.9000
 
+* Adds function `adls_append()` to append a file.
+
 * Adds functions `adls_open_to_text()` and `adls_open_to_file()` to wrap/augment `adls_open_to_raw()`.
 
-* Adds function `adls_open_to_raw()` to read file and return raw data.
+* Adds function `adls_open_to_raw()` to read a file and return raw data.
 
 * Adds function `adls_rename()` to rename a file or directory.